refactoring dependencies
